flashbench: fix building on 32 bits
This fixes all warnings and one confusion of off_t and size_t

+++ dev.h
@@ -7,7 +7,7 @@ struct device {
 	void *readbuf;
 	void *writebuf[3];
 	int fd;
-	ssize_t size;
+	off_t size;
 };
 
 enum writebuf {
